+++ src/vue/forms/DepositForm.vue
@@ -99,6 +99,10 @@ export default {
   },
   mixins: [FormMixin],
 
+  props: {
+    assetCode: { type: String, default: '' },
+  },
+
   data () {
     return {
       CoinpaymentsDepositModule,
@@ -129,7 +133,8 @@ export default {
         .filter(item => item.isDepositable)
 
       if (this.assets.length) {
-        this.selectedAsset = this.assets[0]
+        this.selectedAsset = this.assets.find(a => a.code === this.assetCode) ||
+          this.assets[0]
       }
       this.isLoaded = true
     } catch (e) {